+++ src/main/java/com/rodrigopeleias/myresources/service/UserManagerService.java
@@ -13,10 +13,10 @@ public interface UserManagerService {
 	User getUserByUsername(String username);
 	User getUserById(long id);
 	User getUserByUsernameAndPassword(String username, String password);
-	void removeUser(int id);void addRole(Role role);
+	void removeUser(long id);void addRole(Role role);
 	void updateRole(Role role);
 	Role getRoleByName(String rolename);
 	List<Role> listRoles();
-	void removeRole(int id);
+	void removeRole(long id);
 
 }
